mode-changer/pyproject.toml
[project]
name = "mode-changer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "M-Kashif-ALI", email = "158289645+M-Kashif-ALI@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.9",
    "python-dotenv>=1.1.1",
]

[project.scripts]
mode-changer = "mode_changer:main"
task = "mode_changer.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
